Create auto-encoder class
Changes
- Created the EncodeData class which will allow us to train a
convolutional auto-encoder and then get a lower dimensional
representation of the image data to use for the similarity measure.
Additionally I have created a temporary script that allows us to work
directly with this object. Eventually we will combine all three of
the classes: EncodeData, SimilarityMeasure, and LocalSearch into one
coherent script so we go can from start to finish all in one place.