Update local search formulation
Changes
- Changed formulation from computing the minimum pair-wise distances
to minimizing the sum of variances. We made this change because
the old formulation was simply not working and we wanted to go
with something that had a better chance of a success. As a
consequence of this decision there have been a number of changed
to the directory structure which we detail below

- We removed all similarity measure and old local search scripts.
We no longer need to compute the pairwise similarity distances
as this is not how we compute the objective function. These
have been moved to the old scripts directory

- We removed the encode_data script since we are no longer using
a convolutional auto-encoder to transform the image data. Instead
we are now using a pre-trained model from Keras to transform
the images

- We removed image_distance because this is not needed and now
how we're computing image similarity anymore

- We added the scripts local_search_variance and random_assignment.
The former details the new approach we are taking to solve the
problem and the latter is a benchmark we can use to compare
model performance (among other methods)

- We added the label_embedding_tree script. This is another benchmark
we will use which can be found in the literature

- We removed the reduce_labels.jl file. We are no longer using
the MIP approach to solve the problem and JuMP does not support
MINLP (to the best of my knowledge; it might, but I don't know how
to use it) and so it is no longer a relevant script

- We updated the transform_data to account for many possible pre-trained
models that we could use so that we can give ourselves as many
options as possible.

- Added the benchmark_methods.jl. This script allows us to benchmark
the performance of a given label map assuming that we know the
underlying label distribution

- Removing kmeans.py since it is too short of a script to be useful;
it might come back at a later point.