Update code directory
Changes
1. To simplify our code base, we have combined into to major elements:
"core" and "experiments." Core represents the general scripts we call
for our methodology and experiments corresponds to different
experiments we have used to verify our approach

2. Removed benchmark_methods.jl, local_search_variance.jl,
random_assignment.jl, and label_embedding_tree.py because for the
three .jl files, we are migrating to a mostly python code base for
simplicity, and we removed label_embedding_tree because it has been
subsumed by hc.py in "core"

3. Update transform_data.py to simplify how we infer the label of a
given file under some assumptions regarding data structure and
we no longer need to compute the running mean because we do
normalization in hc.py

4. Added all of the scripts we used for the FMoW experiments where
we transform the data, test a HC versus other benchmarks, and then
compute the relevant evaluation metrics and determine the inferred
meta-classes. Additionally, we visualize the results using the
two .R files: validation_res and vis_bootstrap_res. In the first file
we look more a the algorithmic performance of our estimator and in the
second we focus mainly on the bootstrapped distributions

5. Add bootstrap_results.R to core which is a general set of functions
in R which allows us to generate a bootstrap distribution from our
experiment results.

6. Add kmeans.py which is a short script which defines our kmeans-mean
heuristic we developed for grouping labels

7. Add db2.py which implements the "Divide-by-2" heuristic for grouping
labels

8. Add cnn.py which is a set of functions to build a CNN under our
hierarchical framework (this might be removed later because it might be
subsumed by the hc.py script, but I don't want to delete it just yet
until I can determine its utility more fully)

9. I will now go file-by-file in the FMoW experiments to discuss
briefly what each one does
    - check_img_size.py: Determines the median image size to get an
    appropriate value for re-sizing the images

    - compute_metrics.py: Computes the metrics we are interested in
    evaluating; TODO: we need to generalize this code and add it to
    core so we can easily replicate this for future experiments

    - crop_imgs.py: Takes the starting training and validation data
    and crops them as the creators of the FMoW data did

    - get_inferred_metaclasses.py: Uses the experimental results to
    determine the meta-classes that were inferred by our method

    - get_transformed_data.py: Transforms the training data into a new
    feature space to be used by our method

    - hc_test.py: Large script that utilizes the code for hc.py and
    runs an experiment (this can likely be cleaned up so that we can
    more easily replicate the methodology for future experiments)