Add coordinate descent heuristic
Changes
1. Changed the .gitignore to ignore Cython generated files (relates
to the code we wrote for the coordinate descent heuristic) and
got rid of things that we don't need to ignore anymore

2. Add coordinate descent code in the .pyx code so that we can call
it with quick Cython code and a wrapper for Python so that it is
accessible to other users

3. Put all of the label grouping work into a single script --
group_labels.py. This allows the user to pass an argument for which
label grouping algorithm to use and the code will call the appropriate
script. This will allow us to get ride of the k_means.py script because
it is a very simple function and thus not necessary to have its own
script. We do need the coordinate descent heuristic to have its own
script because it's fairly complex

4. Updated the hc_test to allow us to specify which grouping heuristic
to use for the problem. This also updates the experiment settings
DataFrame so that we can track this later down the road