Create R package and change core files
Changes
1. Updated transform_data to allow for the possibility that none of
the images need to be shrunk and thus only need to be transformed. Also
updated the documentation to correct the claim that we were only using
the Xception model

2. Update .gitignore to no longer ignore the .Rproj file per the
recommendation of the RStudio team

3. Remove bootstrap_results because it has been replaced with a
Python file which does the same thing

4. Replace the old FMoW files of compute_metrics and hc_test and now
only have test_hc which does the same thing except in a cleaner way

5. Add Python scripts which perform the skin cancer experiments

6. Add R package getBootResults which takes in a bootstrapped DataFrame
and generates LaTeX code to make tables and .pdf plots to display the
histograms. The package can be called from other scripts allowing the
R code to be portable and generalizable.