Update community detection
Changes
1. Added ability to plot the time comparison and to compare
Node Top 1 values. These are in flux, but I want to push them and
change them later

2. Updated the way we can make queries when deleting experiments
to allow regular expressions

3. Added the new definition for NT1 in gen_boot_distns; this is also
in flux, but again pushing to keep up to date with changes

4. Added multiple similarity/distance metrics to the community
detection approach

5. Added the "metrics" argument to run_model.py to account for the
changes we've made for the CD approach

6. Added the R scripts we've been using to make the new bootstrap
distributions, training time comparisons, and NT1 comparison. The
old scripts will be deprecated in future iterations